% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{add_cyclic_components}
\alias{add_cyclic_components}
\title{Add Cyclic Time Components to Dataframe}
\usage{
add_cyclic_components(data, time_col = "tstamp")
}
\arguments{
\item{data}{A dataframe that contains a timestamp of type POSIXct or POSIXt}

\item{time_col}{The name of the timestamp column (as a string)}
}
\value{
The original dataframe with qsin and qcos appended
}
\description{
Add Cyclic Time Components to Dataframe
}
\examples{
my_data <- data.frame(tstamp = as.POSIXct("2026-01-01 12:00:00", tz = "UTC"))
add_cyclic_components(my_data, time_col = "tstamp")

}
