% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{add_date}
\alias{add_date}
\title{Add a Date column to the input data frame}
\usage{
add_date(data, time_col, tz = "UTC")
}
\arguments{
\item{data}{A data frame that contains a timestamp}

\item{time_col}{The name of the timestamp column (as a string)}

\item{tz}{A character vector that represents the timezone}
}
\value{
The original data frame with a date column appended
}
\description{
Add a Date column to the input data frame
}
\examples{
data(solar)
solar <- add_date(solar, "tstamp")

}
