% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{add_weights}
\alias{add_weights}
\title{Add solar geometry weights to a dataframe}
\usage{
add_weights(
  df,
  tstamp_col,
  beta,
  period,
  lat,
  lon,
  method = "gaussian",
  tz = "UTC"
)
}
\arguments{
\item{df}{A dataframe.}

\item{tstamp_col}{The name of the timestamp column.}

\item{beta}{The exponential decay parameter.}

\item{period}{An integer that represents the periodicity of the time series in minutes}

\item{lat}{Latitude of the location.}

\item{lon}{Longitude of the location.}

\item{method}{A character vector to specify the weight calculation methodology (gaussian, laplacian)}

\item{tz}{A character vector to specify timezone}
}
\description{
Add solar geometry weights to a dataframe
}
