% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_performance_report}
\alias{generate_performance_report}
\title{Generate a report given a model and a test set.}
\usage{
generate_performance_report(
  test_df,
  model,
  format = "json",
  metric_list = list(rmse = rmse, nmae = nmae)
)
}
\arguments{
\item{test_df}{A data farme used for testing the model}

\item{model}{A instnace of class SolarModel}

\item{format}{A string; the fomrat of the report to be generated. Avialable values: "raw", "json", "data.frame". The default is "raw" i.e. a list is returned.}

\item{metric_list}{A list of named metric functions, from the available metrics}
}
\value{
A report in a specified format
}
\description{
Generate a report given a model and a test set.
}
