% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_distance_from_solar_noon}
\alias{get_distance_from_solar_noon}
\title{Calculate a vector of weights between 0 and 1 for each interval.
The weight represents the relative importance a value has on a specific time interval within the day
and are calculated as the exponential difference between the current timestamp and the}
\usage{
get_distance_from_solar_noon(
  date_time,
  beta = 0.01,
  period = 15,
  method = "gaussian",
  lat = 37.98381,
  lon = 23.727539,
  tz = "UTC"
)
}
\arguments{
\item{date_time}{A POSIXct/POSIXt vector of timestamps}

\item{beta}{A real number that dictates the steepness of the exponental function. The larger the beta the more aggressively the values will decay from the position of the solar noon interval.}

\item{period}{An integer that represents the periodicity of the time series in minutes}

\item{method}{A character vector to specify the weight calculation methodology (gaussian, laplacian)}

\item{lat}{A real number to specify geographical latitude}

\item{lon}{A real number to specify geographical longitude}

\item{tz}{A character vector to specify timezone}
}
\value{
A vector of real numbers between 0 and 1, and their sum is equal to 1.
}
\description{
Calculate a vector of weights between 0 and 1 for each interval.
The weight represents the relative importance a value has on a specific time interval within the day
and are calculated as the exponential difference between the current timestamp and the
}
\examples{
get_distance_from_solar_noon(date_time = lubridate::ymd_hms("2026-01-03 22:00:00"), period = 15)

}
