% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_interval}
\alias{get_interval}
\title{Calculate the position of an interval within a day}
\usage{
get_interval(date_time, period_in_minutes)
}
\arguments{
\item{date_time}{A POSIXct/POSIXt vector of timestamps}

\item{period_in_minutes}{An integer that represents the duration of the interval in minutes}
}
\value{
A vector of integers where each value represents the position of the corresponding timestamp within the day
}
\description{
Calculate the position of an interval within a day
}
\examples{
get_interval(lubridate::ymd_hms("2026-01-03 22:00:00"), 15)

}
