% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_month}
\alias{get_month}
\title{Extract the month from a timestamp}
\usage{
get_month(date_time)
}
\arguments{
\item{date_time}{A POSIXct/POSIXt vector of timestamps}
}
\value{
An integer value between 1 and 12
}
\description{
Extract the month from a timestamp
}
\examples{
get_month(as.POSIXct("2026-01-03 22:00:00", tz = "UTC"))

}
