% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_solar_noon}
\alias{get_solar_noon}
\title{Calculate the timestamp of the solar noon for a given date and geographical postion}
\usage{
get_solar_noon(date_time, lat = 37.98381, lon = 23.727539, tz = "UTC")
}
\arguments{
\item{date_time}{A POSIXct/POSIXt vector of timestamps}

\item{lat}{A real number to specify geographical latitude}

\item{lon}{A real number to specify geographical longitude}

\item{tz}{A character vector to specify timezone}
}
\value{
A data frame with two columns: the date and the corresponding interval that is calculated by the function
}
\description{
Calculate the timestamp of the solar noon for a given date and geographical postion
}
\examples{
get_solar_noon(as.Date("2026-03-01"))

}
