% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_sunrise_sunset}
\alias{get_sunrise_sunset}
\title{Calculate sunrise and sunset for a specific geographical location and timezone}
\usage{
get_sunrise_sunset(date_time, lat = 37.98381, lon = 23.727539, tz = "UTC")
}
\arguments{
\item{date_time}{A POSIXct/POSIXt vector of timestamps}

\item{lat}{A real number that represents latitude}

\item{lon}{A real number that represents longitude}

\item{tz}{A character vector that represents the timezone}
}
\value{
A data frame with dates, sunrise, and sunset timestamps
}
\description{
Calculate sunrise and sunset for a specific geographical location and timezone
}
\examples{
get_sunrise_sunset(as.POSIXct("2026-02-02 23:00", tz = "UTC"))

}
