% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{is_daylight}
\alias{is_daylight}
\title{Calculate Daylight Flag}
\usage{
is_daylight(date_time, lat = 37.98381, lon = 23.727539, tz = "UTC")
}
\arguments{
\item{date_time}{A POSIXct/POSIXt vector of timestamps}

\item{lat}{A numeric value representing latitude (default is Athens)}

\item{lon}{A numeric value representing longitude (default is Athens)}

\item{tz}{A string representing the timezone}
}
\value{
A logical vector (TRUE if during daylight, FALSE otherwise)
}
\description{
Indicates whether a specific timestamp falls between sunrise and sunset.
}
\examples{
#Assuming tstamp is a vector of POSIXct times
data(solar)
solar <- solar |> dplyr::mutate(is_sunny = is_daylight(tstamp))

}
