% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mae}
\alias{mae}
\title{Calculates the Mean Absolute Error (MAE)}
\usage{
mae(y, yhat)
}
\arguments{
\item{y}{A numeric vector of target values.}

\item{yhat}{A numeric vector of predicted values.}
}
\value{
A numeric value
}
\description{
Calculates the Mean Absolute Error (MAE)
}
