% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mape}
\alias{mape}
\title{Calculates the Mean Absolute Percentage Error (MAPE)}
\usage{
mape(y, yhat)
}
\arguments{
\item{y}{A numeric vector of target values.}

\item{yhat}{A numeric vector of predicted values.}
}
\value{
A numeric value
}
\description{
Note: Use with caution for solar data, as true zeros (nighttime)
will produce infinite values. Consider using nMAE instead.
}
