% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ninja}
\alias{ninja}
\title{Renewables Ninja Benchmark Data}
\format{
A data frame with columns:
\describe{
\item{time}{Timestamp}
\item{local_time}{Timestamp}
\item{electricity}{Output in MW}
\item{irradiance_direct}{Direct radiation (W/m2)}
\item{irradiance_diffuse}{Diffuse radiation (W/m2)}
\item{temperature}{Ambient temperature (C)}
}
}
\source{
\url{https://www.renewables.ninja/}
}
\usage{
ninja
}
\description{
Hourly PV and weather data for a 2000 MW simulated plant.
}
\keyword{datasets}
