% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{nmae}
\alias{nmae}
\title{Normalized Mean Absolute Error (nMAE) Factory}
\usage{
nmae(capacity)
}
\arguments{
\item{capacity}{A numeric value representing the max capacity of the plant.}
}
\value{
A function that accepts two numeric vectors, \code{y} (targets) and
\code{yhat} (predictions), and returns the nMAE normalized by the pre-configured
capacity.
}
\description{
Returns a function configured with a specific plant capacity for evaluation.
}
\examples{
solar_nmae <- nmae(capacity = 500)
targets <- c(100, 200, 300)
predictions <- c(110, 190, 310)
solar_nmae(targets, predictions)

}
