% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ppredict}
\alias{ppredict}
\title{Predict solar production with automatic night-time zeroing}
\usage{
ppredict(
  model,
  beta,
  method,
  newdata,
  tstamp_col,
  period,
  lat,
  lon,
  tz = "UTC"
)
}
\arguments{
\item{model}{A trained SVM model.}

\item{beta}{The optimized beta value used during training.}

\item{method}{A character vector "laplacian" or "gaussian"}

\item{newdata}{A data frame that contains unseen data.}

\item{tstamp_col}{The string name of the timestamp column.}

\item{period}{An integer period in minutes}

\item{lat}{A numeric value representing latitude.}

\item{lon}{A numeric value representing longitude.}

\item{tz}{String representing the timezone (default "UTC").}
}
\description{
Predict solar production with automatic night-time zeroing
}
