% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prune_by_weight}
\alias{prune_by_weight}
\title{Sample training data based on physical weights}
\usage{
prune_by_weight(data)
}
\arguments{
\item{data}{The training dataframe containing a 'weight' column}
}
\value{
A pruned dataframe
}
\description{
Sample training data based on physical weights
}
