% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmse}
\alias{rmse}
\title{Calculates the Root Mean Squared Error (RMSE)}
\usage{
rmse(y, yhat)
}
\arguments{
\item{y}{A numeric vector of target values (truth).}

\item{yhat}{A numeric vector of predicted values (estimates).}
}
\value{
A numeric value in the same units as y
}
\description{
Calculates the Root Mean Squared Error (RMSE)
}
