% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rse}
\alias{rse}
\title{Calculates the Relative Squared Error (RSE)}
\usage{
rse(y, yhat)
}
\arguments{
\item{y}{A numeric vector of target values.}

\item{yhat}{A numeric vector of predicted values.}
}
\value{
A numeric value representing the ratio of model error to baseline error
}
\description{
Calculates the Relative Squared Error (RSE)
}
