% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solar}
\alias{solar}
\title{National Solar Production Data}
\format{
A data frame with timestamps and PV production.
}
\usage{
solar
}
\description{
15-minute interval data for the Greek national grid.
}
\keyword{datasets}
