% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_model.R
\name{SolarModel}
\alias{SolarModel}
\title{R6 Class for Solar Forecasting}
\description{
A high-level interface for training, predicting, and evaluating solar power models.
Encapsulates plant metadata, importance-weighting logic, and consistent metrics.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lat}}{Numeric. Latitude of the territory or plant.}

\item{\code{lon}}{Numeric. Longitude of the territory or plant.}

\item{\code{tz}}{String. Timezone (e.g., "Europe/Athens" or "UTC").}

\item{\code{capacity}}{Numeric. Installed capacity used for cnMAE normalization.}

\item{\code{period}}{Integer. Time series frequency in minutes (default: 15).}

\item{\code{beta}}{Numeric. The exponential decay parameter for importance weights.}

\item{\code{method}}{String. "laplacian" or "gaussian".}

\item{\code{model}}{The trained model object (typically from e1071).}

\item{\code{metrics}}{List. A named list of metric functions.}

\item{\code{tstamp_col}}{String. The column name of the timestamp.}

\item{\code{cal_df}}{Data frame. A data frame used for conformal predictions.}

\item{\code{q_hat}}{Numeric. The quantile estimation used for calculating prediction intervals using conformal statistics.}

\item{\code{model_formula}}{R formula. The formula passed to fit function to train the model.}

\item{\code{scaling_params}}{A list that contains the scale parameters if kernlab is used. As opposed to e1071 kernlab does nbot automatically handle the parameter scaling.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{mode}}{Access or set the current operational mode ("train" or "eval").}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SolarModel-new}{\code{SolarModel$new()}}
\item \href{#method-SolarModel-prepare_data}{\code{SolarModel$prepare_data()}}
\item \href{#method-SolarModel-fit}{\code{SolarModel$fit()}}
\item \href{#method-SolarModel-predict}{\code{SolarModel$predict()}}
\item \href{#method-SolarModel-save_model}{\code{SolarModel$save_model()}}
\item \href{#method-SolarModel-clone}{\code{SolarModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-SolarModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SolarModel object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarModel$new(
  lat,
  lon,
  capacity,
  period,
  tz = "UTC",
  engine = NULL,
  metrics = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{A numeric value representing latitude.}

\item{\code{lon}}{A numeric value representing longitude.}

\item{\code{capacity}}{A numeric value; the installed nameplate capacity.}

\item{\code{period}}{An integer; the period of the time series in minutes (default: NULL).}

\item{\code{tz}}{A string representing the timezone (default: "UTC").}

\item{\code{engine}}{The SVM engine to use ("e1071" or "kernlab").}

\item{\code{metrics}}{An optional named list of metric functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarModel-prepare_data"></a>}}
\if{latex}{\out{\hypertarget{method-SolarModel-prepare_data}{}}}
\subsection{Method \code{prepare_data()}}{
Helper to prepare features and optionally split data.
Automatically switches behavior based on the object's current mode.
The data returned depends on the current mode of the model. If the mode is train
train_test_split function is called.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarModel$prepare_data(
  df,
  tstamp_col,
  model_formula,
  prop = 0.8,
  prop_cal = 0,
  method = "laplacian",
  beta = 0.01,
  add_weights = FALSE,
  detect_outliers = TRUE,
  drop_outliers = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{Data frame to process.}

\item{\code{tstamp_col}}{Character. The name of the timestamp column in \code{df}.}

\item{\code{model_formula}}{An object of class \code{formula} describing the model.}

\item{\code{prop}}{Numeric. The proportion of data for the training set (default: 0.8).}

\item{\code{prop_cal}}{Numeric. The proportion of data for the calibration set (default: 0).}

\item{\code{method}}{Character. The weighting method, either "laplacian" or "gaussian".}

\item{\code{beta}}{Numeric. The decay parameter for importance weights (default: 0.01).}

\item{\code{add_weights}}{Logial. Add weights to data frame rows.}

\item{\code{detect_outliers}}{Logical. If set to TRUE a one class SVM will be trained, will detect outliers and place a flag at each row in the training set.}

\item{\code{drop_outliers}}{Logical. If set to TRUE the outliers will be removed from the training set.}

\item{\code{...}}{Additional arguments: \code{perc_conf} (percentage confidence) if a one class svm is used for detecting outliers
where \code{perc_conf} expresses the level of confidence (percentage) with which we would like the model to label a data point as being an outlier.
and \code{col_names}, a character vector of column names from the training data frame to be passed in the one-class SVM.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
In "train" mode, a list with \code{train_df} and \code{test_df}. In "eval" mode, a processed data frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-SolarModel-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the model using the internal training engine.
Automatically chooses between standard training and beta co-optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarModel$fit(train_df, test_df, parallel = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{train_df}}{Data frame containing the training data.}

\item{\code{test_df}}{Data frame containing the validation/test data.}

\item{\code{parallel}}{A logical value indicating whether a custom parallel tuning should be used}

\item{\code{...}}{Optional arguments:
\itemize{
\item \code{n_folds}: N integer, the number of folds to use in parallel tuning
\item \code{ctrl}: A list of control parameters for the tuning process.
\item \code{train_grid}: A data frame or list defining the SVM hyperparameter
search space.
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing \code{metrics} (test set performance) and
\code{yhat} (predictions on the test dataset).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarModel-predict"></a>}}
\if{latex}{\out{\hypertarget{method-SolarModel-predict}{}}}
\subsection{Method \code{predict()}}{
Generate physical-floor-clipped predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarModel$predict(new_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{Data frame containing features for prediction.}

\item{\code{tstamp_col}}{Character. The name of the timestamp column in \code{new_data}.}

\item{\code{newdata}}{Data frame for which to generate predictions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with columns \code{date_time} and \code{yhat}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarModel-save_model"></a>}}
\if{latex}{\out{\hypertarget{method-SolarModel-save_model}{}}}
\subsection{Method \code{save_model()}}{
Save the current model instance to an RDS file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarModel$save_model(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Character string. The path where the model should be saved.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolarModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SolarModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
