% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{train_model}
\alias{train_model}
\title{Train, tune, and evaluate an SVM model using the e1071 package}
\usage{
train_model(
  train_df,
  test_df,
  model_formula,
  metrics_list,
  ctrl = NULL,
  train_grid = NULL
)
}
\arguments{
\item{train_df}{Data frame used to train the model.}

\item{test_df}{Data frame used to evaluate the model (must have same features).}

\item{model_formula}{Formula of type \code{target ~ covariates}.}

\item{metrics_list}{A named list of functions used for evaluation (e.g., list(rmse = rmse, mae = mae)).}

\item{ctrl}{SVM tune control.}

\item{train_grid}{SVM hyperparameter grid.}
}
\value{
A list with the model, hyperparameters, and test metrics.
}
\description{
Train, tune, and evaluate an SVM model using the e1071 package
}
