% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{train_oc_model}
\alias{train_oc_model}
\title{Detect outliers in the training set
The critical parameter in this function is perc_conf represents the confidence with which a pattern in the input data frame
is charactetized as being an outlier. Specifying a high confidence level is an expression of preference to true positives with the trade-off that some
outliers will remain undetected.}
\usage{
train_oc_model(data_frame, perc_conf = NULL)
}
\arguments{
\item{data_frame}{The data frame that contains the training data}

\item{perc_conf}{An integer close to (but less than) 100}
}
\value{
A vector of logical values
}
\description{
Detect outliers in the training set
The critical parameter in this function is perc_conf represents the confidence with which a pattern in the input data frame
is charactetized as being an outlier. Specifying a high confidence level is an expression of preference to true positives with the trade-off that some
outliers will remain undetected.
}
\examples{
col_names <- c("pv", "irrad", "temp", "clds")
train_oc_model(solar[, col_names])

}
