% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{train_test_split}
\alias{train_test_split}
\title{Stratified Train-Test Split for Solar Time-Series}
\usage{
train_test_split(df, tstamp_col, period_in_minutes, prop = 0.8, prop_cal = 0)
}
\arguments{
\item{df}{A dataframe containing solar data}

\item{tstamp_col}{Character vector; The string name of the timestamp column}

\item{period_in_minutes}{Integer; The interval resolution (e.g., 15 or 60)}

\item{prop}{Real number between 0 and 1; The proportion of data to include in the training set}

\item{prop_cal}{Real number between 0 and 1; the proportion of data to include in the calibration set}
}
\value{
A list of data frames that contains \code{train} and \code{test} and \code{cal}
}
\description{
Stratified Train-Test Split for Solar Time-Series
}
