% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{wmae}
\alias{wmae}
\title{Calculates the Weighted Mean Absoulte Error}
\usage{
wmae(y, yhat, weight)
}
\arguments{
\item{y}{A numeric vector of target values.}

\item{yhat}{A numeric vector of predicted values.}

\item{weight}{A numeric vector of weight values.}
}
\value{
A numeric value
}
\description{
Calculates the Weighted Mean Absoulte Error
}
